"This should take like, 2-3 hours or so." -Me, 2-3 days ago.
TL:DR I made another child class of Object that creates spheres of a given radius,
and teselation level. So we can control how many polygons it has. Seemed like a
good idea at the time but we'll see. But anyhow I have it so you can specify if you
want the norms to be the triangle normals (A.K.A. have the sphere appear edged)
or you can have it smooth (use barycentric interpolate between verts to create
smooth lighting.)

I'm pressently working on having Objects take textures to use instead of colors.
More to come, loving this project. (Beauty is pain)