I have learned a valueable lesson about trying to "free-hand" changes without making a commit first. That was a bit of a hassel.
I added a ton of new ways to mess with the sphere, like toggleing the smooth vs flat shading methods and adding random colors and whatnot.

In addition, probably more intersting, I made a significant change so that we have Vertex objects, and can set up the Objects per-vertex info with a vector of those instead of a massive GLfloat array I resize myself.

I also followed along with video 11 I believe and added a "Mesh" class which is just a lamer version of the Object class. Regardless, it was a stepping stone to adding in the Vertex thing so it was good.

More to come.